Added different drawing options for edges. Currently the option to use is hard-coded. Need to make this toggle-enabled with GUI. Also, need to update bounding rectangle appropriately
git-svn-id: file:///home/jleigh/var/svn-shared/Hapnet/src/trunk@147 dfa6d6af-2265-4a09-bd68-ba6c179305d8

+++ gui/NetworkLayout.cpp
@@ -49,7 +49,7 @@ void NetworkLayout::mapEdges()
       EdgeData ed;
       ed.start = i;
       ed.end = child.data(NetworkItem::EdgeEndRole).toInt();
-      ed.prefLength = NetworkItem::EDGELENGTH * child.data(NetworkItem::SizeRole).toInt();//SCALEFACTOR * _model->data(child, NetworkItem::SizeRole).toInt();
+      ed.prefLength = NetworkItem::EDGELENGTH * child.data(NetworkItem::SizeRole).toDouble();//SCALEFACTOR * _model->data(child, NetworkItem::SizeRole).toInt();
       _edgeList.push_back(ed);
             
     }
